requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
dj-rest-auth==2.1.10
Django==3.1.6
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
Faker==8.12.0
filelock==3.0.12
geojson==2.5.0
idna==2.9
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mixer==7.1.2
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.10.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyowm==3.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.4.2
wcwidth==0.2.5
